re-ran the notebook
